Issue #1014: Move getExtensionPaths from kernel to modules/extensions
Moves extension path resolution from kernel to modules/extensions for
architectural compliance with CLAUDE.md lazy-loading requirements.

Changes:
- Move kernel/internal/extension-paths.ts to modules/extensions/internal/
- Update kernel/index.ts to remove export
- Update modules/extensions/index.ts to add exports
- Update voice extension import to use modules/extensions
- Move test file to tests/unit/modules/extensions/
- Update all READMEs to reflect new location

The kernel module should only contain "always-needed" code, but
getExtensionPaths() is only called when extensions are loaded.
Moving it to modules/extensions follows the lazy-loading architecture
and keeps extension-related utilities together.
