+++ kernel/README.md
@@ -31,7 +31,6 @@ kernel/
     async-queue.ts
     config.ts
     defaults.ts
-    extension-paths.ts
     logger.ts
     errors.ts
     lru-map.ts
@@ -64,27 +63,3 @@ const registry = new PluginRegistry('./plugins');
 // Optional fail-fast validation for long-lived processes
 await registry.validateAll();
 ```
-
-### Extension Paths
-Extensions can use `getExtensionPaths()` to resolve their plugin directories reliably:
-
-```ts
-import { getExtensionPaths } from '@/kernel/index.ts';
-
-// Resolves extension root and plugins directory
-const { extensionRoot, pluginsRoot } = getExtensionPaths('voice');
-
-// Results are cached for performance (cache invalidates when cwd changes)
-```
-
-**Resolution order** (first match with existing plugins directory wins):
-1. External roots from `llm-adapter.paths.json` (if configured via `paths.lookup.extensions.externalRoots`)
-2. `PACKAGE_ROOT/extensions/{extensionId}/plugins` (production/dist)
-3. `cwd/extensions/{extensionId}/plugins` (development/source fallback)
-
-**Features:**
-- Integrates with `getAdapterPathsConfig()` for external root support
-- Cache invalidates automatically when `process.cwd()` changes
-- Paths are canonicalized via `realpathSync` to handle symlinks (e.g., macOS `/var` → `/private/var`)
-
-This handles the dist/source path discrepancy where TypeScript compiles to `dist/` but JSON manifests remain in source.
