+++ modules/extensions/README.md
@@ -30,3 +30,27 @@ Extensions may ship a `extensions/<name>/defaults.json`.
 This file is intended to be loaded only when an extension is enabled/invoked, and can be merged with
 legacy configuration overlays provided by the server.
 
+## Extension path resolution
+
+Extensions should use `getExtensionPaths()` to resolve their plugin directories reliably:
+
+```ts
+import { getExtensionPaths } from '@/modules/extensions/index.js';
+
+// Resolves extension root and plugins directory
+const { extensionRoot, pluginsRoot } = getExtensionPaths('voice');
+
+// Results are cached for performance (cache invalidates when cwd changes)
+```
+
+**Resolution order** (first match with existing plugins directory wins):
+1. External roots from `llm-adapter.paths.json` (if configured via `paths.lookup.extensions.externalRoots`)
+2. `PACKAGE_ROOT/extensions/{extensionId}/plugins` (production/dist)
+3. `cwd/extensions/{extensionId}/plugins` (development/source fallback)
+
+**Features:**
+- Integrates with `getAdapterPathsConfig()` for external root support
+- Cache invalidates automatically when `process.cwd()` changes
+- Paths are canonicalized via `realpathSync` to handle symlinks (e.g., macOS `/var` → `/private/var`)
+
+This handles the dist/source path discrepancy where TypeScript compiles to `dist/` but JSON manifests remain in source.
