+++ modules/extensions/index.ts
@@ -9,3 +9,13 @@ export {
   loadExtensionDefaults,
   resolveExtensionEntry
 } from './internal/extension-lookup.js';
+
+export type {
+  ExtensionPaths,
+  GetExtensionPathsOptions
+} from './internal/extension-paths.js';
+
+export {
+  getExtensionPaths,
+  resetExtensionPathsCache
+} from './internal/extension-paths.js';
